kinship sparse

Compute a sparse kinship matrix from pedigree data, a pedigree object or a pedigreeList object

Usage

kinship_sparse(id, ...)

# S3 method for default
kinship_sparse(id, dadid, momid, sex, chrtype = "autosome", ...)

# S3 method for pedigree
kinship_sparse(id, chrtype = "autosome", ...)

# S3 method for pedigreeList
kinship_sparse(id, chrtype = "autosome", ...)

Arguments

id

either a pedigree object, pedigreeList object, or a vector of subject identifiers. Subject identifiers may be numeric or character.

...

extra arguments to pass forward to internal functions.

dadid

for each subject, the identifier of the biological father. This is only used if id is a vector.

momid

for each subject, the identifier of the biological mother. This is only used if id is a vector.

sex

vector of sex values coded as 1=male, 2=female

chrtype

chromosome type. The currently supported types are "autosome" and "X" or "x".

Details

This is very minor modification of the kinship2::kinship() function. The only difference is that the kinship matrix returned by this function is always a sparse matrix. This can be usefull when working with large interconnected pedigrees, which cannot be cut into smaller pedigrees by the kinship2::makefamid function.

References

Sinnwell J, P, Therneau T, M, Schaid D, J: The kinship2 R Package for Pedigree Data. Hum Hered 2014;78:91-93. doi: 10.1159/000363105

Examples

library(kinship2)
#> Loading required package: Matrix
#> Loading required package: quadprog
data(minnbreast)
mb <- minnbreast
bpeds <- pedigree(mb$id, mb$fatherid, mb$motherid, 
                  mb$sex, affected=mb$proband, 
         famid=mb$famid)

identical(as.matrix(kinship_sparse(bpeds[1])),kinship(bpeds[1]))  
#> [1] TRUE

mb2 <- minnbreast[minnbreast$famid==4,]

identical(
   kinship(mb2$id,mb2$fatherid,mb2$motherid,mb2$sex),
   as.matrix(kinship_sparse(mb2$id, mb2$fatherid, mb2$motherid, mb2$sex))
)
#> [1] TRUE

# Note that these are not very efficient:
mb3 <- minnbreast[minnbreast$famid %in% 4:50,]
system.time(kinship_sparse(mb3$id, mb3$fatherid, mb3$motherid, mb3$sex))
#>    user  system elapsed 
#>   7.419   0.093   7.512 
system.time(kinship_sparse(pedigree(mb3$id, mb3$fatherid, mb3$motherid, mb3$sex)))
#>    user  system elapsed 
#>   3.251   0.029   3.279 

# Splitting into families and computing per per family is much faster: 
system.time(makefamid(mb3$id, mb3$fatherid, mb3$motherid))
#>    user  system elapsed 
#>   0.050   0.000   0.049 
system.time(kinship(pedigree(mb3$id, mb3$fatherid, mb3$motherid, mb3$sex, famid=mb3$famid)))
#>    user  system elapsed 
#>   0.060   0.000   0.061