PA-FGRS

Compute posterior mean and variance of liability of index person given disease status, thresholds and proportion of risk observed in relatives.

Usage

pa_fgrs(
  rel_status,
  thr = NA,
  rel_thr = rep(thr, length(rel_status)),
  rel_w = rep(1, length(rel_status)),
  covmat,
  i_status = NA,
  i_thr = thr,
  i_w = 1,
  conditional.mix = T
)

Arguments

rel_status

vector with status of relatives 1=affected, 0=not affected, NA= missing

thr

threshold value used for all individuals (relatives and index person)

rel_thr

optional vector of specific threshold for each relative

rel_w

optional vector of proportion of risk experienced by each relative

covmat

covariance matrix of liabilities. First row and column correspoding to index person,subsequent rows and columns corresponding to relatives.

i_status

optional status of index person, if to be considered 1=affected, 0=not affected, NA= missing

i_thr

optional specific threshold for index person

i_w

optional proportion of risk experienced by index person

conditional.mix

logical indicator of wether the mixture of proportions should be conditional (default=TRUE).

Examples

pa_fgrs(c(0,1),qnorm(.9),covmat = matrix(c(.5,.25,.25,.25,1,.25,.25,.25,1),3))
#>     postM   postVar 
#> 0.3695726 0.4311061